Add a setting to override the package version
A generate package version actually depends on the version of the
OpenAPI spec and on the version of the generator. Upgrading the
generator version may break backward compatibility of the client. This
gives the user the possibility to bump its client version to reflect
changes in the generated code even though the spec has not changed.