fix: run commands in package directory if meta=none
`project_dir` is set as the working directory if meta generation is
turned off. Commands cwd to `project_dir`, so hooks will be run in
whatever directory you're running the `generate` command in.
The default post-hooks will then run recursively in that directory
on any Python files it can find.

Set the directory to run commands in to `package_dir` if meta is none,
so post-hooks only touch the generated code.

fixes #696