Union variants use Title names instead of indices
Changed behavior from using the term `x_type_0` to `x_type_y` if x is a Schema that defines a Title property string `y`.

This provides an easier interface for understanding what users are importing, as well as less reliance on the ordering of the schema.

BREAKING CHANGE: The name of imports will change if the Schema they were built on previously used Title properties on fields within a oneOf Union.