validate value of 'const' properties
For 'const' properties, the parsed value will now be checked against the
value in the schema at runtime and a ValueError will be raised if they
do not match.

This helps with parsing `oneOf` when matching a const property is the
only way to resolve which of the types should be returned.