better support for merging properties with allOf (#1096)
Fixes #1090 (with some limitations - see below)

This PR allows the generator to work with some valid schemas where the
same property name is specified by more than one schema within an
`allOf`, where previously it would have failed.

# Background

As described in the issue, OpenAPI and JSON Schema define `allOf` in a
fairly vague and tolerant way. If a schema is defined like this:
```yaml
  C:
    allOf:
      - $ref: "/path/to/schema/A" 
      - $ref: "/path/to/schema/B"
```

—then the validation for C is "must pass validation for A, and for B". A
and B are both objects, and they both have property X, then the value
for X must pass validation for A.X and B.X; the validation is the
strictest subset of both.

This does not map very cleanly to the client generator's internal model.
Because its ultimate goal is to producing Python model classes with
properties of specific types, we need to be able to "merge" A.X and B.X
to produce a single instance of some ___Property class, whose attributes
are derived from both.

The logic for doing this was narrower than necessary. If, for instance,
A.X and B.X only differed in terms of their `description`, or if one of
them specified a `default` value and the other did not, the generator
would reject them as being too different.

# Solution

This PR defines a more nuanced property merge logic, as follows:

- If both versions of the property are of the same type, the result is a
property of the same type. Its validation rules, if any, are the
strictest subset of the validation rules for both (for instance, "string
with max length 3" and "string with max length 5" produces "string with
max length 3").
- Int and number (aka float) properties can always be merged; the result
is an int property (since "must be integer" is a stricter rule than
"must be a number"). This is only invalid if there is a non-integer
default value.
- "Any" can be merged with any property type; the result is that
property type (since "must be a ___" is stricter than "can be
anything").
- For common string-valued attributes such as `description`, earlier
values are overridden by later values, on the assumption that the second
schema is meant to be an extension or specialization of the first.

It also preserves helpful aspects of the original implementation:

- Two enums can be merged, if the values for one are a subset of the
other; the result uses the subset.
- A string enum can be merged with a string, or an int enum with an int;
the result is an enum.

# Limitations

Due to limitations of our internal model, some combinations of
properties that are valid in OpenAPI will still be rejected by this
implementation:

- Two string properties with different `pattern` attributes cannot be
merged, even if there are values that could match both of those regexes.
- Two list types that contain different schemas cannot be merged, even
though logically this should mean "every element in the list must match
both of those schemas".
- OpenAPI allows `minimum` and `maximum` to be specified for numeric
types, but (as far as I can tell) the generator currently ignores these
altogether, so the merge logic doesn't take them into account either.

# Compatibility

This PR preserves the previous schema parsing behavior in all cases
where the generator would have previously succeeded. The new behaviors
all apply to valid schemas where the generator would have previously
failed.

---------

Co-authored-by: Dylan Anthony <dbanty@users.noreply.github.com>
Co-authored-by: Dylan Anthony <43723790+dbanty@users.noreply.github.com>