fix: import cast for required const properties, since it's used in the template (#1153)
The
[template](https://github.com/openapi-generators/openapi-python-client/blob/main/openapi_python_client/templates/property_templates/const_property.py.jinja#L2)
for `const` properties uses `cast` unconditionally, but the import for
`cast` in `ConstProperty.get_imports` was conditioned on whether the
property was required or not, leading to broken generated code for
required const properties.

I'd be happy to add a test for this if desired! A cursory glance
suggests that maybe the end_to_end tests would be the right place?

Resolves #1150