Support multiple tags in each endpoint (#687)
Currently when an endpoint has multiple tags, the first tag is used and
everything else is ignored.

This PR modifies it so endpoints with multiple tags are added to each of
the tags.

Yes, this results in repeated code :sweat_smile:, but works beautifully
and functions can now be found anywhere we expect them to be.

---------

Co-authored-by: Dylan Anthony <dbanty@users.noreply.github.com>
Co-authored-by: Dylan Anthony <43723790+dbanty@users.noreply.github.com>