test: Rename python to python3 in tests (#1223)
This commit renames the invocation of `python` to `python3` in the
tests.

While many systems now alias `python` to `python3`, this behavior is not
guaranteed across all environments and is often configured by the user,
rather than provided by the operating system directly.

Explicitly using `python3` ensures that the tests are executed with the
intended Python 3 interpreter, regardless of user-specific
configurations,
and therefore improves the portability and reliability of the test
suite.