pass models & enums as lists so custom templates can use them (#1189)
Resolves #1188.

The change is simply to store a list, rather than a generator
expression.

The new `models_init.py.jinja` template in the end-to-end tests shows
that this now works. Without the fix in `openapi.py`, this template is
rendered with nothing under the `Using "openapi.models"` and `Using
"openapi.enums"` headings.

This fix is somewhat important for our use case, because we need to put
some code in `models/__init__.py` that is aware of what the model class
names and modules are. Without the fix, the only way to get this would
be to iterate over `imports`, which gives you strings like `from
.module_name import ClassName`, and then parse those strings, which is
workable but pretty clunky.