feat: add handling of endpoint default response
This modifies the internal Response object to use either `HTTPStatus`
or a string literal "default" for status codes and modifies parsing
in the `Endpoint` class to match.

This additionally modifies the endpoint jinja2 template to handle
this alternate response status code field by using the default response,
if it exists, as the fallback case instead of the UnexpectedStatus error
or returning None, and modifies the return typing to match.